Fix deadlock bug in epoll proactor.
The issue is about main io loop blocking on epoll_wait when in fact there are active fibers that could be running.
It's either dispatch fiber that has not run yet or user fibers that were activated but the state machine has not
took it into the account. Now I wrote enough comments that hopefully further refactorings won't introduce
this bug again. In addition, I added the test to epoll_proactor_test that reproduced the bug with
some probability.

Signed-off-by: Roman Gershman <romange@gmail.com>