feat(TlsSocket): add ownership semantics to TlsSocket (#98)
* feat(TlsSocket): add ownership semantics to TlsSocket

TlsSocket is constructed by a pointer to the next socket,
effectively acting as a decorator/proxy object of the
underline socket. This however is problematic, because
once a connection upgrades to tls, you now need to keep alive
two different objects with whose lifetime semantics are bound.
Therefore, we make TlsSocket own the underline socket, to avoid
decoupling the lifetime of the two sockets.