An attempt fixing the bug with inconsistent state in fibers (#131)
* An attempt fixing the bug with inconsistent state in fibers

Specifically, fix #130.
The code that crashed makes sure that a fiber is not in remote_ready queue
when it wakes up: it yields if it is, and thus allows the dispatcher to
empty the remote queue and pull the active fiber from there.

The code crashes because after the Yield call, it appears that the fiber is still
in the remote queue. Based on the code inspection, it seems unprobable that
a dispatcher won't empty the remote queue, therefore I suspect that remote queue
was corrupted in the first place.

This fix focuses on ScheduleFromRemote (producer sider) that before that did not protect against concurrent calls
for the same fiber. It was partly justified because ScheduleFromRemote is usually called from the
WaitQueue code under lock. However, there are other cases that can call this function,
and maybe they corrupted the queue.

In addition, we removed the epoch based notifications which did not contribute to the correctness.

Signed-off-by: Roman Gershman <romange@gmail.com>

* Update util/fibers/detail/scheduler.cc

Co-authored-by: Roy Jacobson <roi.jacobson1@gmail.com>
Signed-off-by: Roman Gershman <romange@gmail.com>

---------

Signed-off-by: Roman Gershman <romange@gmail.com>
Co-authored-by: Roy Jacobson <roi.jacobson1@gmail.com>