chore: reimplement fiber migrate
Before, we used a temporary fiber to pass along the migrated fiber to another proactor.
Now, we simplify the transition by using the dispatcher fiber that also handles I/O loop.
We switch to it and execute the delegation to another proactor.

Also, remove the old code and get rid of sccache that is not very reliable.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>