chore: add support for direct file descriptors
Provide the infrastructure for handling direct file descriptors in iouring.
Those descriptors bypass the usual linux tables and can be handled only by iouring API.
For now we provide a very small application for them by using them with listen sockets,
just to get some mileage with the code before we push it to more use-cases.

Signed-off-by: Roman Gershman <romange@gmail.com>
Signed-off-by: Roman Gershman <roman@dragonflydb.io>