chore: general improvements in iuring/fibers
The main change is that we now rely on liburing function io_uring_peek_batch_cqe
to peek for completions instead of manually iterating with `io_uring_for_each_cqe` call.
The reason for this is that iouring API evolved, and the logic for peeking completions became
non-trivial: it requires calling a system call in some rare cases. So
in terms correctness and ease of maintainance it's better to rely on the library function to do the right job.

In addition, fibers scheduler code has non-functional changes:
a. Added a trace variable to ease with debugging: it shows the last transitioned state of the fiber
   in ambigous situations.
b. ProcessSleep now returns the number of woken fibers.

Finally, we improve listen-socket code in accept server to work better with iouring socket during
error returning scenarios.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>