fix: mpsc intrusive queue now reports its precise state during Pop operation (#215)
Before, we could get "no element" response when in fact the queue is not empty.
This can happen when head is being in the middle of the update but its "next" field
has not been updated yet.

Before the change, we treated this state as "Empty", while the queue could have multiple
entries. After this change: we retry the pop operation in scheduler
until we reach the true depletion of the remote queue.

Also, clean up a debug variable that I used to chase this bug.

Signed-off-by: Roman Gershman <romange@gmail.com>