chore: Fix the crash in tls shutdown
The root cause was due to a socket being closed and destroyed inside ListenerInterface::RunSingleConnection
and in parallel the listener was performing the shutdown sequence inside ListenerInterface::RunAcceptLoop().

In fact we had a comment saying it's working because socket.Shutdown did not preempt.
In one of my recent PRs I implemented the logic where TlsSocket informs the peer about the shutdown
which made the call preemptible. As a result the whole management code around this became unsafe.

1. I introduced an intrusive refcnt so that a connection would not be destroyed inside RunSingleConnection
   when RunAcceptLoop still uses it.
2. Changed the iteration loop in the latter tu support preemption.
3. Fixed the Shutdown code to support the case where the tcp socket is closed in the middle of shutdown.
4. Found unrelated migration bug that where we failed to Unlink a connection because the listener was destroyed.
   Now we try to unlink only if the connection indeed is linked inside a connection list.

Signed-off-by: Roman Gershman <romange@gmail.com>