feat: introduce FiberPriority
Also, perform some clean-up in the code i touched.
FiberPriority is not used yet, it only stored in FiberInterface.
The only functional change is that it fibers use boost fixedsize_stack, we now pass the stack size through the Opts struct,
which is 64KB, while default boost fixedsize_stack size was 128KB, hence the change in fibers_test.cc

Signed-off-by: Roman Gershman <romange@gmail.com>