chore: add high priority to the scheduler (#457)
High priority share the NORMAL queue with NORMAL fibers.

If a high priority fiber is activated, it is pushed to the front of the queue.
If it yields, however, it is still pushed to back.

Also added support for priorities for default, main thread that does not run an I/O polling loop.
It is mainly added for being able to test fibers without proactors.

Signed-off-by: Roman Gershman <romange@gmail.com>