epoll_socket: Notify fibers on shutdown (#475)
A conditional wake after shutdown for epoll socket is converted to
unconditional wait.

When a socket shuts down, any fibers reading from it must get a timely
event to be able to react and exit their loops. If the socket is closed
before this happens, the socket fd is removed from the epoll interest
set, and the callback handler is reset. This can result in the reading
fibers being suspended forever causing a hang.

To avoid this a wake is issued which allows such fibers to terminate
their receive loop.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>