chore: integrate EnableRecvMultishot in uring socket with OnRecv hook (#482)
* chore: integrate EnableRecvMultishot in uring socket with OnRecv hook

Before we had a monstrous and inefficient interface,trying to juggle
recv multishot notifications together with Recv calls. This was akward, inconvenient and
required multiple tracking layers.

With `OnRecv` hook, we can just pass the provided buffer to the registered callback as we
receive it and immediately replenish it. Bundles processing also becomes much simpler though
we still do not support them in this PR.

I have not added tests yet but checked manually via echo_server.

Signed-off-by: Roman Gershman <romange@gmail.com>

* chore: address comments

Signed-off-by: Roman Gershman <romange@gmail.com>

---------

Signed-off-by: Roman Gershman <romange@gmail.com>