chore: stop yielding inside tls socket
Before: we yielded in tls socket in case both read and write path are stuch on the same I/O operation.
This is problematic, when this I/O op does not progress. Imagine a case where a client does not read data, and
the server is stuck on writing to the socket. If the read path asks to write to the socket, and it yields,
it will just constantly yield spiking to 100% CPU on that thread as long as the write path does not progress.

This PR removes yields and replaces them with simple blocking condition.

Signed-off-by: Roman Gershman <romange@gmail.com>