fix: crash-bug if we deregister onrecv hook too quickly (#501)
EpollAdd was asynchronous while EpollDel is synchronous.
It is possible that we ask to register EpollAdd but before it actually happenned
we try to deregister in the kernel but kernel does not know this id yet.

This PR calls io_submit inside EpollAdd which flashes the SQEs
and ensures they are executed immediately.

Signed-off-by: Roman Gershman <romange@gmail.com>